Refactor state management by moving StateNotification class to a new file and updating its integration in the extension
