+++ src/extension.ts
@@ -1,6 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import { StateNotification } from './stateNotification';
 
 // Global variables to store language configuration disposable and document state
 let languageConfigDisposable: vscode.Disposable | undefined;
@@ -244,48 +245,3 @@ export function deactivate() {
     documentListeners = [];
 }
 
-class StateNotification {
-    
-    private _statusBarItem: vscode.StatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
-    private _isEnabled: boolean = true;
-    
-    public updateState(enabled: boolean) {
-        this._isEnabled = enabled;
-        
-        if (enabled) {
-            this.enabled();
-        } else {
-            this.disabled();
-        }
-    }
-    
-    public enabled() {
-        // Show notification
-        vscode.window.showInformationMessage('SongTxt mode is now enabled');
-        
-        // Update the status bar
-        this._statusBarItem.text = '$(music) SongTxt: Enabled';
-        this._statusBarItem.tooltip = 'SongTxt syntax highlighting is active. Click to disable.';
-        this._statusBarItem.command = 'extension.disableSongtxt';
-        this._statusBarItem.show();
-    }
-
-    public disabled() {
-        // Show notification
-        vscode.window.showInformationMessage('SongTxt mode is now disabled');
-        
-        // Update the status bar
-        this._statusBarItem.text = '$(circle-slash) SongTxt: Disabled';
-        this._statusBarItem.tooltip = 'SongTxt syntax highlighting is disabled. Click to enable.';
-        this._statusBarItem.command = 'extension.enableSongtxt';
-        this._statusBarItem.show();
-    }
-
-    public get isEnabled(): boolean {
-        return this._isEnabled;
-    }
-
-    dispose() {
-        this._statusBarItem.dispose();
-    }
-}
